Restructured Node Controller and AWS Layers to be Clearer on Separation
