+++ node-video-crime-miner/src/AWS Layer/Rekognition/videoClient.ts
