+++ node-video-crime-miner/src/AWS Layer/Rekognition/videoUtils.ts
