+++ node-video-crime-miner/src/AWS Layer/s3Connector.ts
