+++ node-video-crime-miner/src/AWS Layer/snsClient.ts
