+++ node-video-crime-miner/src/AWS Layer/sqsClient.ts
