+++ node-video-crime-miner/src/Node Connector Layer/rekognitionConnector.ts
